youtube-upload-pipeline/requirements.txt
# YouTube Upload Pipeline Dependencies

# Google API and Authentication
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.88.0

# Image processing for thumbnails
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Video processing
moviepy>=1.0.3

# Data handling
pandas>=2.0.0
requests>=2.28.0

# Configuration and utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
click>=8.1.0

# Existing integrations
httpx>=0.25.0
asyncio-throttle>=1.0.2

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0